feat(backup): rich progress reporting

+++ clifs/plugins/backup/backup.py
@@ -4,12 +4,35 @@ import csv
 import shutil
 import time
 from argparse import ArgumentParser, Namespace
-from collections import Counter
 from pathlib import Path
-from typing import List, NamedTuple, Optional, Tuple
+from typing import Dict, Iterable, List, NamedTuple, Optional, Tuple
+
+from rich.console import Console, RenderableType
+from rich.live import Live
+from rich.panel import Panel
+from rich.progress import (
+    BarColumn,
+    Progress,
+    TaskID,
+    TaskProgressColumn,
+    TimeRemainingColumn,
+)
+from rich.table import Table
+from rich.theme import Theme
 
 from clifs import ClifsPlugin
-from clifs.utils_cli import cli_bar, print_line
+from clifs.utils_cli import print_line
+
+THEME_RICH = Theme(
+    {
+        "bar.complete": "default",
+        "bar.finished": "green",
+        "bar.back": "bright_black",
+        "progress.percentage": "default",
+        "progress.remaining": "bright_black",
+    },
+)
+CONSOLE = Console(theme=THEME_RICH)
 
 
 class DirPair(NamedTuple):
@@ -19,23 +42,23 @@ class DirPair(NamedTuple):
     dest: Path
 
 
-def conditional_copy(path_source: Path, path_dest: Path, dry_run: bool = False) -> int:
+def conditional_copy(
+    path_source: Path, path_dest: Path, dry_run: bool = False
+) -> Optional[str]:
     """
     Copy only if dest file does not exist or is older than the source file.
     """
     process = None
     if not path_dest.exists():
-        process = "copying"
+        process = "adding"
     elif (path_source.stat().st_mtime - path_dest.stat().st_mtime) > 1:
-        process = "updating from"
+        process = "updating"
 
     if process is not None:
-        print(f" - {process} {str(path_source)}", flush=True)
         if not dry_run:
             path_dest.parent.mkdir(parents=True, exist_ok=True)
             shutil.copy2(path_source, path_dest)
-        return 1
-    return 0
+    return process
 
 
 def conditional_delete(
@@ -44,14 +67,11 @@ def conditional_delete(
     """
     Delete only if `path_source` is not in `list_source`.
     """
-    if path_source not in list_source:
+    if path_source not in list_source and path_dest.exists():
         if path_dest.is_dir():
-            print(f" - deleting dir {str(path_dest)}", flush=True)
             if not dry_run:
                 shutil.rmtree(str(path_dest))
         else:
-            filename = path_dest.name
-            print(f" - deleting {filename}", flush=True)
             if not dry_run:
                 path_dest.unlink()
         return 1
@@ -73,6 +93,21 @@ def list_filedirs(dir_source: Path) -> Tuple[List[Path], List[Path]]:
             list_files.append(cur_file)
 
     return list_files, list_dirs
+
+
+class LastActionProgress(Progress):
+    """Progress showing the last action in a separate line."""
+
+    def get_renderables(self) -> Iterable[RenderableType]:
+        """Get a number of renderables for the progress display."""
+
+        table = self.make_tasks_table(self.tasks)
+        yield table
+        for task in self.tasks:
+            yield (
+                f"{task.fields.get('last_action_desc','Last action')}: "
+                f"{task.fields.get('last_action', '-')}"
+            )
 
 
 class FileSaver(ClifsPlugin):
@@ -159,60 +194,199 @@ class FileSaver(ClifsPlugin):
         time_start = time.time()
 
         for dir_pair in self.dir_pairs:
-            self.backup_dir(
-                dir_pair.source, dir_pair.dest, delete=self.delete, dry_run=self.dry_run
-            )
+            self.backup_dir(dir_pair.source, dir_pair.dest)
         time_end = time.time()
         time_run = (time_end - time_start) / 60
         print(f"Hurray! All files backed up in only {time_run:5.2f} minutes")
 
-    @staticmethod
     def backup_dir(
+        self,
         dir_source: Path,
         dir_dest: Path,
-        delete: bool = False,
-        dry_run: bool = False,
     ) -> None:
-        print(f"Backing up files from:\n{dir_source}\nin:\n{dir_dest}.")
         print_line()
-        list_files_source, list_dirs_source = list_filedirs(dir_source)
-        # initialize stats
-        counter = Counter(checked=len(list_files_source))
+        print(f"Backing up files from:\n    {dir_source}\nin:\n    {dir_dest}.")
 
-        for cur_num_file, cur_file in enumerate(list_files_source, start=1):
-            counter["copied"] += conditional_copy(
-                cur_file,
-                Path(str(cur_file).replace(str(dir_source), str(dir_dest))),
-                dry_run=dry_run,
-            )
-            cli_bar(cur_num_file, counter["checked"], suffix="of files checked")
-        print_line()
-        if delete:
+        files_source, dirs_source = list_filedirs(dir_source)
+        self.copy_data(
+            dir_source=dir_source,
+            dir_dest=dir_dest,
+            files_source=files_source,
+        )
+
+        if self.delete:
             print("All files stored, checking for files to delete now.")
-            print_line()
-            list_files_dest, list_dirs_dest = list_filedirs(dir_dest)
 
-            num_dest = len(list_files_dest)
-            for cur_num_file, cur_file_dest in enumerate(list_files_dest, start=1):
-                counter["deleted"] += conditional_delete(
+            self.delete_obsolete_data(
+                dir_source=dir_source,
+                dir_dest=dir_dest,
+                files_source=files_source,
+                dirs_source=dirs_source,
+            )
+        print_line()
+
+    def copy_data(self, dir_source: Path, dir_dest: Path, files_source: List[Path]):
+        progress: Dict[str, Progress] = {
+            "counts": Progress(
+                "{task.description}",
+                "{task.completed}",
+            ),
+            "overall": LastActionProgress(
+                "{task.description}",
+                BarColumn(),
+                TaskProgressColumn(),
+                TimeRemainingColumn(),
+            ),
+        }
+
+        tasks: Dict[str, TaskID] = {
+            "progress_backup": progress["overall"].add_task(
+                "Storing data:  ", total=len(files_source), last_action="-"
+            ),
+            "count_files_found": progress["counts"].add_task(
+                "Files processed:", total=None
+            ),
+            "count_files_added": progress["counts"].add_task(
+                "Files added:", total=None
+            ),
+            "count_files_updated": progress["counts"].add_task(
+                "Files updated:", total=None
+            ),
+            "count_files_untouched": progress["counts"].add_task(
+                "Files untouched:", total=None
+            ),
+        }
+
+        progress_table = Table.grid()
+        progress_table.add_row(
+            Panel.fit(
+                progress["overall"],
+                title="Progress Backup",
+                border_style="cyan",
+                padding=(2, 2),
+            ),
+            Panel.fit(
+                progress["counts"],
+                title="Counts",
+                border_style="bright_black",
+                padding=(1, 2),
+            ),
+        )
+        with Live(progress_table, refresh_per_second=5, console=CONSOLE):
+            for cur_file in files_source:
+                progress["counts"].advance(tasks["count_files_found"])
+                action = conditional_copy(
+                    cur_file,
+                    Path(str(cur_file).replace(str(dir_source), str(dir_dest))),
+                    dry_run=self.dry_run,
+                )
+                if action:
+                    progress["overall"].update(
+                        tasks["progress_backup"],
+                        last_action=f"{action} {cur_file.name}",
+                    )
+                    if action == "adding":
+                        progress["counts"].advance(tasks["count_files_added"])
+                    elif action == "updating":
+                        progress["counts"].advance(tasks["count_files_updated"])
+                else:
+                    progress["counts"].advance(tasks["count_files_untouched"])
+                progress["overall"].advance(tasks["progress_backup"])
+
+    def delete_obsolete_data(
+        self,
+        dir_source: Path,
+        dir_dest: Path,
+        files_source: List[Path],
+        dirs_source: List[Path],
+    ):
+        files_dest, dirs_dest = list_filedirs(dir_dest)
+
+        progress: Dict[str, Progress] = {
+            "counts": Progress(
+                "{task.description}",
+                "{task.completed}",
+            ),
+            "overall": LastActionProgress(
+                "{task.description}",
+                BarColumn(style="green"),
+                TaskProgressColumn(),
+                TimeRemainingColumn(),
+            ),
+        }
+
+        tasks_delete: Dict[str, TaskID] = {
+            "progress_delete_files": progress["overall"].add_task(
+                "Deleting files:",
+                total=len(files_dest),
+                last_action="-",
+                last_action_desc="Last file deleted",
+            ),
+            "progress_delete_folders": progress["overall"].add_task(
+                "Deleting dirs:",
+                total=len(dirs_dest),
+                last_action="-",
+                last_action_desc="Last dir deleted",
+            ),
+            "count_files_found": progress["counts"].add_task(
+                "Files processed:", total=None
+            ),
+            "count_files_deleted": progress["counts"].add_task(
+                "Files deleted:", total=None
+            ),
+            "count_folders_found": progress["counts"].add_task(
+                "Dirs processed:", total=None
+            ),
+            "count_folders_deleted": progress["counts"].add_task(
+                "Dirs deleted:", total=None
+            ),
+        }
+
+        progress_table = Table.grid()
+        progress_table.add_row(
+            Panel.fit(
+                progress["overall"],
+                title="Progress Deletion",
+                border_style="magenta",
+                padding=(1, 2),
+            ),
+            Panel.fit(
+                progress["counts"],
+                title="Counts",
+                border_style="bright_black",
+                padding=(1, 2),
+            ),
+        )
+
+        with Live(progress_table, refresh_per_second=5, console=CONSOLE):
+            for cur_file_dest in files_dest:
+                progress["counts"].advance(tasks_delete["count_files_found"])
+                action = conditional_delete(
                     Path(str(cur_file_dest).replace(str(dir_dest), str(dir_source))),
                     cur_file_dest,
-                    list_files_source,
-                    dry_run=dry_run,
+                    files_source,
+                    dry_run=self.dry_run,
                 )
-                cli_bar(cur_num_file, num_dest, suffix="of files checked")
+                if action:
+                    progress["counts"].advance(tasks_delete["count_files_deleted"])
+                    progress["overall"].update(
+                        tasks_delete["progress_delete_files"],
+                        last_action=cur_file_dest.name,
+                    )
+                progress["overall"].advance(tasks_delete["progress_delete_files"])
 
-            for cur_dir_dest in list_dirs_dest:
-                conditional_delete(
+            for cur_dir_dest in dirs_dest:
+                progress["counts"].advance(tasks_delete["count_folders_found"])
+                action = conditional_delete(
                     Path(str(cur_dir_dest).replace(str(dir_dest), str(dir_source))),
                     cur_dir_dest,
-                    list_dirs_source,
-                    dry_run=dry_run,
+                    dirs_source,
+                    dry_run=self.dry_run,
                 )
-            print_line()
-
-        print(
-            f"\nStored {counter['copied']} files out of {counter['checked']} from "
-            f"'{dir_source}'.\n"
-            f"Deleted {counter['deleted']} files in destination directory."
-        )
+                if action:
+                    progress["counts"].advance(tasks_delete["count_folders_deleted"])
+                    progress["overall"].update(
+                        tasks_delete["progress_delete_folders"],
+                        last_action=cur_dir_dest.name,
+                    )
+                progress["overall"].advance(tasks_delete["progress_delete_folders"])
